+++ src/python/approxeng/task/__init__.py
@@ -9,20 +9,20 @@ RESOURCES = {}
 LOG = logging.getLogger('approxeng.task')
 
 
-def task(_func=None, *, name=None, resources=None):
+def task(_func=None, *, name=None):
     """
     Decorator to indicate that a function is a simple task. The function will be registered, using either the
     name if explicitly provided, or the name of the function otherwise.
     """
     if _func is not None:
         # Called with no name argument
-        register_task(name=_func.__name__, value=_func, resources=None)
+        register_task(name=_func.__name__, value=_func)
         return _func
     else:
         # Called with an explicit argument, use this to register it
         def decorator(func):
             task_name = name if name is not None else func.__name__
-            register_task(name=task_name, value=func, resources=resources)
+            register_task(name=task_name, value=func)
             return func
 
         return decorator
@@ -55,13 +55,17 @@ class Task(ABC):
     Abstract base class for tasks, things which are called repeatedly to perform some higher function.
     """
 
+    global_count = 0
+
     class World:
         """
         The world that the task tick sees, consists of all resources the task said it needed.
         """
 
-        def __init__(self, resources):
-            self._dict = {task_resource: RESOURCES[task_resource].value() for task_resource in resources}
+        def __init__(self, resources, global_count, task_count):
+            self.dict = {task_resource: RESOURCES[task_resource].value() for task_resource in resources}
+            self.dict['global_count'] = global_count
+            self.dict['task_count'] = task_count
 
         def __getitem__(self, item):
             if isinstance(item, tuple):
@@ -69,12 +73,12 @@ class Task(ABC):
             return self.__getattr__(item)
 
         def __getattr__(self, item: str):
-            if item in self._dict:
-                return self._dict[item]
+            if item in self.dict:
+                return self.dict[item]
             raise AttributeError
 
         def __contains__(self, item):
-            return item in self._dict
+            return item in self.dict
 
     def __init__(self, name, resources=None):
         """
@@ -93,6 +97,7 @@ class Task(ABC):
             self._resources = [resources]
         self.active = False
         self.name = name
+        self.task_count = 0
 
     @property
     def resources(self):
@@ -131,14 +136,18 @@ class Task(ABC):
                 RESOURCES[task_resource].shutdown()
             self.active = False
 
-    def do_tick(self, count):
+    def do_tick(self):
         """
         Start up the task, if needed, then call the tick method, passing in the world and tick count.
         """
         if not self.active:
             self.do_startup()
-        LOG.debug('Task "%s" tick %i', self.name, count)
-        return self.tick(world=Task.World(self.resources), count=count)
+        LOG.debug('Task "%s", task_tick %i, global_tick %i', self.name, self.task_count, Task.global_count)
+        return_value = self.tick(
+            world=Task.World(resources=self.resources, task_count=self.task_count, global_count=Task.global_count))
+        Task.global_count = Task.global_count + 1
+        self.task_count = self.task_count + 1
+        return return_value
 
     @abstractmethod
     def startup(self):
@@ -155,14 +164,12 @@ class Task(ABC):
         pass
 
     @abstractmethod
-    def tick(self, world, count):
+    def tick(self, world):
         """
         Called every tick, do your stuff here!
 
         :param world:
             A world object, access any required resources through properties on this object.
-        :param count:
-            The global tick count.
         :return:
             Used to signal what to do at the end of the tick:
             None - continue, call the same task again
@@ -186,7 +193,7 @@ class SimpleTask(Task):
     count : monotonically ascending tick count across the entire application.
     """
 
-    def __init__(self, task_function, resources, name):
+    def __init__(self, task_function, name):
         """
         Create a new simple task instance, this is generally going to be called from within the library when wrapping
         a task function.
@@ -200,6 +207,28 @@ class SimpleTask(Task):
             The name of the task, this is only really used internally for logging, the canonical name is defined by the
             key under which the task is registered.
         """
+
+        resources = [key for key in inspect.signature(task_function).parameters.keys()]
+
+        # task_state isn't a normal world resource, we'll explicitly add it later but we don't want the framework
+        # to try to initialise it. Remove it from the list of resources if it's been specified and make a note that
+        # we did so so we know to add it back in later
+        if 'task_state' in resources:
+            resources.remove('task_state')
+            self.uses_state = True
+        else:
+            self.uses_state = False
+
+        # The world always contains task_count and global_count, but we may not want it to. Remember whether these are
+        # used for later.
+        self.use_task_count = 'task_count' in resources
+        self.use_global_count = 'global_count' in resources
+
+        if self.use_task_count:
+            resources.remove('task_count')
+        if self.use_global_count:
+            resources.remove('global_count')
+
         super(SimpleTask, self).__init__(resources=resources, name=name)
         self.task_function = task_function
         self.state = {}
@@ -216,32 +245,29 @@ class SimpleTask(Task):
         """
         self.state.clear()
 
-    def tick(self, world, count):
+    def tick(self, world):
         """
         A single task tick. Calls the task function, supplying it with whatever parameters it needs.
 
         :param world:
             World state, set of resources accessible as properties.
-        :param count:
-            Global tick counter, increases by 1 each time
         :return:
             The return value from the task function. As elsewhere, this is interpreted as follows:
             None - don't change anything, keep this task running, call it again
             True - exit from the task processing loop, shutting down the process
             Task or String - shut this task down, set the named or provided task as the current task
         """
-        signature = inspect.signature(self.task_function)
-        args = {}
-        if 'state' in signature.parameters:
-            args['state'] = self.state
-        if 'world' in signature.parameters:
-            args['world'] = world
-        if 'count' in signature.parameters:
-            args['count'] = count
-        return self.task_function(**args)
+        if self.uses_state:
+            world.dict.update({'task_state': self.state})
+        if not self.use_task_count:
+            del world.dict['task_count']
+        if not self.use_global_count:
+            del world.dict['global_count']
+
+        return self.task_function(**world.dict)
 
 
-def register_task(name, value, resources=None):
+def register_task(name, value):
     """
     Explicitly register a task, either from a function or from an instance of Task
 
@@ -252,15 +278,10 @@ def register_task(name, value, resources=None):
         object. You may want to use the latter, more verbose, form if extensive setup or custom state handling is
         needed by your task, although in general most of such handling should be done with resources and tasks
         themselves should remain largely state free.
-    :param resources:
-        A string, or list of strings, containing the names of resources required for this task to run. These resources
-        will be initialised at the same time as the task itself, and shut down alongside it where appropriate. They are
-        accessible from the world object passed to the tick function. This is not used if a Task object is passed, as
-        these already specify their resource lists.
     """
     if isinstance(value, types.FunctionType):
-        TASKS[name] = SimpleTask(name=name, task_function=value, resources=resources)
-        LOG.info('Registered task function "%s", required resources: %s', name, resources)
+        TASKS[name] = SimpleTask(name=name, task_function=value)
+        LOG.info('Registered task function "%s", required resources: %s', name, TASKS[name].resources)
     elif isinstance(value, Task):
         TASKS[name] = value
         LOG.info('Registered task class "%s", required resources: %s', name, value.resources)
@@ -279,14 +300,27 @@ class Resource(ABC):
 
     @abstractmethod
     def startup(self):
+        """
+        Called before the resource is used by any tasks which require it. It should never be called when already active,
+        but code defensively here and check.
+        """
         pass
 
     @abstractmethod
     def shutdown(self):
+        """
+        Called after a task which used this resource has finished. Also called if another resource is registered with
+        the same name, or if the task loop exits. You must implement this to handle multiple calls sensibly, i.e. check
+        that you haven't already tidied up, don't assume this will only be called after the resource has been started.
+        """
         pass
 
     @abstractmethod
     def value(self):
+        """
+        Get the value of the resource. The result of this will be passed to any tasks through the world object, or to
+        any task functions requesting the resource through a named parameter.
+        """
         pass
 
 
@@ -320,7 +354,9 @@ def register_resource(name, value):
     resource class instance. If it is already a resource class instance just register it. If it's a plain static value
     then wrap it in a function that always returns that value, then wrap that up in the simple class.
     """
-
+    if name in RESOURCES:
+        # If this resource was already defined we're going to overwrite it, so shut the existing one down first
+        RESOURCES[name].shutdown()
     if isinstance(value, types.FunctionType):
         RESOURCES[name] = SimpleResource(name=name, value_func=value)
         LOG.info('Registered resource function "%s"', name)
@@ -335,30 +371,27 @@ def register_resource(name, value):
         LOG.info('Registered resource value "%s"', name)
 
 
-class ExitTask(Task):
+@task(name='exit')
+def exit_task(error=None):
     """
-    Switch control to this task to exit the task loop. Registered with the name 'exit' for convenience
+    Exit the loop. If there was an exception raised causing this task to run then wrap the exception in the TaskStop,
+    otherwise it'll be empty.
+    """
+    return TaskStop(error)
+
+
+class TaskStop:
+    """
+    Wraps a single value, defaulting to None. If a task returns an instance of this class, the task loop will exit and
+    the return value of the run() function will be the wrapped value. This allows tasks to exit the loop and pass
+    information to the caller as they do so.
     """
 
-    def __init__(self):
-        super(ExitTask, self).__init__(name='exit')
-
-    def startup(self):
-        pass
-
-    def shutdown(self):
-        pass
-
-    def tick(self, world, count):
-        # At this point we've got access to the error message through world.error
-        return True
+    def __init__(self, return_value=None):
+        self.return_value = return_value
 
 
-# Register the exit task as 'exit'
-register_task(name='exit', value=ExitTask())
-
-
-def run(root_task, error_task='exit', check_tasks=None):
+def run(root_task, error_task='exit', check_tasks=None, raise_exceptions=False):
     """
     Run the task loop!
 
@@ -368,13 +401,20 @@ def run(root_task, error_task='exit', check_tasks=None):
         The task to switch to if an exception occurs, defaults to the ExitTask to exit the loop. If you have hardware
         such as motors that you need to ensure is deactivated, the best approach is to have a task that only does
         hardware shutdown, handles any errors with that process internally, and then delegates to the exit task to
-        stop the task look.
+        stop the task look. Defaults to the exit task if not specified.
     :param check_tasks:
         A sequence of functions which will be called before each tick of the selected task. If any of them return
         then the return value is used instead of calling and using the value of the task's tick. This can be done
         to handle cases like 'make the home button always jump back to the root task', or 'exit the task loop on
         low battery conditions' or similar. Don't put too much logic here, it'll get called every tick. Also good for
         cases where you absolutely want to bail if hardware isn't available (joystick out of range is a particular case)
+    :param raise_exceptions:
+        Defaults to False, if set to True then any exceptions raised by a task will be handled, then wrapped in a
+        TaskException and raised from this call. If False then they will be handled, and control passed to the
+        designated error task.
+    :returns:
+        If the loop exits as the result of a task returning a :class:`~approxeng.task.TaskStop` it will return the
+        value wrapped by that instance, otherwise None.
     """
 
     def get_task(t):
@@ -393,39 +433,54 @@ def run(root_task, error_task='exit', check_tasks=None):
     # Start with the root task as the active one
     active_task = get_task(root_task)
     # Initialise count to 0
-    count = 0
     # Loop until we're done
     finished = False
-    while not finished:
-        try:
-            response = None
-            # If we have any pre-task checks to run do them now. If any of those functions return
-            # non-None values we'll use those in place of the active task. Code these carefully!
-            # Here's where you'd check for e.g. joystick not connected.
-            if check_tasks is not None:
-                for check_task in check_tasks:
-                    check_response = check_task()
-                    if check_response is not None:
-                        response = check_response
-            # If no check_task functions returned anything, run the actual task tick
-            if response is None:
-                response = active_task.do_tick(count)
-            # Increment the tick count
-            count = count + 1
-            # If the tick function returned a value it means we need to switch control
-            if response is not None:
-                if isinstance(response, Task) or isinstance(response, str):
-                    # New task, either name or Task object. Shut down and switch to it for the next tick
-                    active_task.do_shutdown()
-                    active_task = get_task(response)
-                elif isinstance(response, bool) and response is True:
-                    # True value returned, shut the current task down and exit the loop
-                    active_task.do_shutdown()
-                    finished = True
-        except Exception as e:
-            # Anything throwing an exception ends up here. Log it first, then delegate to a handler task
-            LOG.error(e, exc_info=True)
-            # Shut the active task down, add the exception to the world as 'error' and launch the error task
-            active_task.do_shutdown()
-            register_resource('error', e)
-            active_task = get_task(error_task)
+    return_value = None
+    try:
+        while not finished:
+            try:
+                response = None
+                # If we have any pre-task checks to run do them now. If any of those functions return
+                # non-None values we'll use those in place of the active task. Code these carefully!
+                # Here's where you'd check for e.g. joystick not connected.
+                if check_tasks is not None:
+                    for check_task in check_tasks:
+                        check_response = check_task()
+                        if check_response is not None:
+                            response = check_response
+                # If no check_task functions returned anything, run the actual task tick
+                if response is None:
+                    response = active_task.do_tick()
+                # If the tick function returned a value it means we need to switch control
+                if response is not None:
+                    if isinstance(response, Task) or isinstance(response, str):
+                        # New task, either name or Task object. Shut down and switch to it for the next tick
+                        active_task.do_shutdown()
+                        active_task = get_task(response)
+                    elif isinstance(response, TaskStop):
+                        # TaskStop value returned
+                        active_task.do_shutdown()
+                        finished = True
+                        return_value = response.return_value
+            except Exception as e:
+                # Anything throwing an exception ends up here. Log it first, then delegate to a handler task
+                LOG.exception('Exception raised within task look')
+                # Shut the active task down, add the exception to the world as 'error' and launch the error task
+                active_task.do_shutdown()
+                if raise_exceptions:
+                    raise TaskException from e
+                register_resource('error', e)
+                active_task = get_task(error_task)
+    except TaskException as te:
+        # Catch and stash the exception in the return value
+        return_value = te
+    finally:
+        # Finished, shut down all resources and exit
+        for res in RESOURCES:
+            RESOURCES[res].shutdown()
+    # If we're raising exceptions, and there was an exception, raise it.
+    if raise_exceptions and isinstance(return_value, Exception):
+        raise return_value
+    # Otherwise return the return value and exit.
+    return return_value
+